working with uploading video: decomposition
