+++ src/services/studio/file.service.ts
@@ -1,5 +1,5 @@
 import { instance } from '@/api/axios'
-import type { IFileResponse } from '@/types/file.types'
+import type { IFileResponse, IProgressProcessingResponse } from '@/types/file.types'
 
 class FileService {
 	private _UPLOAD_FILE = '/upload-file'
@@ -13,7 +13,9 @@ class FileService {
 	}
 
 	getProcessingStatus(fileName: string) {
-		const data = instance.get<number>(`${this._UPLOAD_FILE}/status/${fileName}`)
+		const data = instance.get<IProgressProcessingResponse>(
+			`${this._UPLOAD_FILE}/status/${fileName}`
+		)
 		return data
 	}
 }
